requirements.txt
discord.py[speed]
poe-api-wrapper
python-dotenv
jishaku